refactor(order): Change to DTO tailored to each request

+++ src/main/java/com/multicampus/gamesungcoding/a11ymarketserver/feature/order/entity/OrderItems.java
@@ -32,7 +32,7 @@ public class OrderItems {
     private String productName;
 
     @Column(nullable = false)
-    private Long productPrice;
+    private Integer productPrice;
 
     @Column(nullable = false)
     private Integer productQuantity;
